taxi: extract optimizer QSim module
